requirements.txt
packaging==23.1
pynput==1.7.6
pyserial==3.5
python-can==4.2.0
pywin32==306
six==1.16.0
typing_extensions==4.5.0
wrapt==1.15.0
